Clean deployContract execute and move functions to helpers
